fix some error

+++ 0x1E-search_algorithms/0-linear.c
@@ -10,12 +10,12 @@
 
 int linear_search(int *array, size_t size, int value)
 {
-	size_t i;
+	int i;
 
 	if (!array)
 		return (-1);
 
-	for (i = 0; i < size; i++)
+	for (i = 0; i < (int) size; i++)
 	{
 		printf("Value checked array[%d] = [%d]\n", i, array[i]);
 		if (array[i] == value)
